refactoring cli classes to improve debugging
